requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
black==22.3.0
certifi==2020.6.20
chardet==3.0.4
click==8.1.3
colorama==0.4.4
defusedxml==0.6.0
dj-database-url==0.5.0
django-allauth==0.42.0
django-ckeditor==5.9.0
django-crispy-forms==2.1
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
djangorestframework-simplejwt==5.2.0
fastapi==0.110.0
greenlet==3.0.3
gunicorn==20.0.4
h11==0.14.0
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.9.0
pillow==10.4.0
platformdirs==2.5.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests-oauthlib==1.3.0
setuptools==71.1.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.2
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2022.1
urllib3==1.25.9
uvicorn==0.29.0
whitenoise==5.1.0
django
djangorestframework
requests>=2.25.0
idna<3.0
